requirements.txt
pygame~=2.1.0
numpy~=1.21.4
PyOpenGL~=3.1.5